+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - You can now click a button in the search bar to search (you can also still just prss enter, of course).
+- You can now filter searches by instance type.
 
 ### Changed